refactor(ts): extract timeline/skillLibrary.ts from store
Move the activeSkillLibrary computed (and its nested buildBaseAction /
buildStandardOrVariantSkill / buildSegmentModels helpers) into a
useSkillLibrary(deps) composable, mirroring the optimizer. Skill-building logic
moved verbatim; pure data helpers are imported directly and the store-local
helpers (cloneJsonData, resolveLevelNumber, getOperatorSkillIcon, etc.) are
injected. Golden + skill-library structure suites unchanged. No behavior change.
